x11-drivers/nvidia-drivers: Add IUSE=egl to install libglvnd EGL control file.
nVidia's closed-source driver does not appear to include a proper
libEGL, instead just reusing the one from libglvnd.  Because of this,
the library will fail to initialize without a libglvnd vendor control
file for the EGL implementation.  USE="-libglvnd egl" will cause this
control file to be installed in /usr/share/glvnd, where libEGL will
find it when /etc/glvnd does not exist (as assumed for the USE=-libglvnd
case).

With USE=libglvnd, USE=egl has no effect and libglvnd control files are
installed in /etc/glvnd as usual.