virtual/opengl: Fix circular dependency on media-libs/libglvnd
Gentoo's virtual/opengl-8 replaces the dependency on Mesa with a
dependency on libglvnd, which is a good sign for upstream eventually
excising explicit libglvnd dependencies from packages that only use
it as an OpenGL check but breaks our workaround for that usage.
We fix the problem by just reverting the change for now, since Mesa
is still required at least for OpenGL header installation.  We could
potentially change to a more flexible dependency if mesa is broken up
into less monolithic packages.